Test suite improvements for usability, correctness and performance (#10864)
NOTE: This was previous opened in #10854 but got closed accidentally.

This does a number of things:
- Use more runners for the integration tests by splitting the tests into groups
- Fix Codecov reporting on Windows
- Separates the documentation build into own workflow (only triggered upon changes to docs files)
- Combine various requirements files.
- Add pytest-rerunfailures to be able to rerun flaky tests.
- Refactor a few unittest style tests to be pytest tests to reduce flakiness.
- Move pytest config file and fixtures in better places.
- Fix temp folder issue on Windows.
- Cleanup Makefile.
- Lint on all platforms, in an own workflow file.

Fix #10846.
