+++ .github/workflows/ci-images.yml
@@ -15,10 +15,10 @@ on:
 
 env:
   REGISTRY: ghcr.io
-  IMAGE_NAME: ${{ github.repository }}
+  IMAGE_NAME: ${{ github.repository }}-ci
 
 jobs:
-  build:
+  linux:
     runs-on: ubuntu-latest
     permissions:
       contents: read
@@ -28,11 +28,6 @@ jobs:
         python: [3.7, 3.8, 3.9]
     steps:
       - uses: actions/checkout@v2
-
-      - name: Lint Dockerfile
-        uses: hadolint/hadolint-action@v1.5.0
-        with:
-          dockerfile: ./ci/Dockerfile
 
       - name: Log in to the Container registry
         uses: docker/login-action@v1
@@ -48,14 +43,14 @@ jobs:
           images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}
           flavor: |
             latest=false
-            suffix=-python${{ matrix.python }}
+            suffix=-linux-python${{ matrix.python }}
 
       - name: Build & push Python ${{ matrix.python }}
         uses: docker/build-push-action@v2
         with:
-          context: ./ci
           file: ./ci/Dockerfile
           build-args: PYTHON_VERSION=${{ matrix.python }}
           labels: ${{ steps.meta.outputs.labels }}
           tags: ${{ steps.meta.outputs.tags }}
-          push: ${{ github.ref == 'refs/heads/master' }}
+          # push: ${{ github.ref == 'refs/heads/master' }}
+          push: true
