+++ .github/workflows/docs.yml
@@ -1,0 +1,28 @@
+name: CI docs
+
+on:
+  push:
+    branches:
+      - master
+    paths:
+      - '.github/workflows/docs.yml'
+      - 'docs/**'
+  pull_request:
+    paths:
+      - '.github/workflows/docs.yml'
+      - 'docs/**'
+
+jobs:
+  docs:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
+      - name: Setup
+        run : |
+          make env-docs
+      - name: Build the docs
+        run : |
+          cd docs
+          conda run --name conda-docs make html
