+++ .github/workflows/lint.yml
@@ -1,0 +1,36 @@
+name: Lint
+
+on:
+  push:
+    branches:
+      - master
+    paths-ignore:
+      - 'docs/**'
+  pull_request:
+    paths-ignore:
+      - 'docs/**'
+
+jobs:
+  lint:
+    strategy:
+      matrix:
+        platform: [ubuntu-latest, macos-latest, windows-latest]
+
+    runs-on: ${{ matrix.platform }}
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - uses: conda-incubator/setup-miniconda@v2
+
+      - name: Setup
+        shell: bash -l {0}
+        run: |
+          conda info
+          conda list
+          conda install flake8
+
+      - name: Lint
+        shell: bash -l {0}
+        run: |
+          python -m flake8 . --statistics
