+++ .readthedocs.yml
@@ -1,5 +1,5 @@
 version: 2
 python:
-  version: 3
+  version: "3"
   install:
   - requirements: docs/requirements.txt
