--- ci/build_docs.sh
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-
-set -o errtrace -o pipefail -o errexit
-
-pip install -r docs/requirements.txt
-cd docs
-make html
