+++ ci/scripts/linux/setup.sh
@@ -1,0 +1,13 @@
+#!/usr/bin/env bash
+
+set -o errtrace -o pipefail -o errexit
+
+apt-get update --fix-missing
+apt-get install -y --no-install-recommends tini wget build-essential bzip2 ca-certificates libglib2.0-0 libxext6 libsm6 libxrender1 git mercurial subversion sudo
+apt-get clean
+rm -rf /var/lib/apt/lists/*
+
+useradd -m -s /bin/bash test_user
+usermod -u 1001 test_user
+groupmod -g 1001 test_user
+echo "test_user ALL=(ALL) NOPASSWD:ALL" >> /etc/sudoers
