+++ ci/scripts/windows/unit.bat
@@ -1,0 +1,5 @@
+@ECHO ON
+pushd %TEMP%
+cd \conda_src
+CALL dev-init.bat
+CALL pytest -m "not integration" -v --splits=%TEST_SPLITS% --group=%TEST_GROUP%
