+++ conda/core/initialize.py
@@ -177,7 +177,6 @@ def initialize_dev(shell, dev_env_prefix=None, conda_source_root=None):
         raise CondaError("Operation failed. Privileged install disallowed for 'conda init --dev'.")
 
     env_vars = {
-        'ADD_COV': '--cov-report xml --cov-report term-missing --cov conda',
         'PYTHONHASHSEED': str(randint(0, 4294967296)),
         'PYTHON_MAJOR_VERSION': python_version[0],
         'TEST_PLATFORM': 'win' if on_win else 'unix',
