--- dev/requirements.txt
@@ -1,6 +1,0 @@
-pycosat >=0.6.3
-pyopenssl >=16.2.0
-requests >=2.12.4,<3
-ruamel_yaml >=0.11.14
-git
-conda-package-handling
