+++ dev/start.bat
@@ -1,12 +1,10 @@
 @IF "%~1"=="" (
-    @SET devenv=%CD%\devenv
+    @SET DEVENV=%CD%\devenv
 ) ELSE (
-    @SET "devenv=%~1"
+    @SET "DEVENV=%~1"
 )
-@IF "%~2"=="" (
-    @SET pyver=3
-) ELSE (
-    @SET "pyver=%~2"
+@IF "%PYTHON%"=="" (
+    @SET "PYTHON=3.8"
 )
 
 @REM Unset some variables that get in the way
@@ -17,17 +15,17 @@ set PYTHONPATH=
 set PYTHONHOME=
 
 @IF EXIST "dev-init.bat" @GOTO :INIT_BUILD
-@IF EXIST "%devenv%\conda-meta\history" @GOTO :ALREADY
-@ECHO Downloading Miniconda%pyver%-latest-Windows-x86_64.exe as miniconda.exe
-@powershell.exe -NoProfile -Command (new-object System.Net.WebClient).DownloadFile('https://repo.continuum.io/miniconda/Miniconda%pyver%-latest-Windows-x86_64.exe','miniconda.exe')
-@ECHO Installing miniconda to: %devenv%
-@start /wait "" miniconda.exe /InstallationType=JustMe /RegisterPython=0 /AddToPath=0 /S /D=%devenv%
-call "%devenv%"\Scripts\activate.bat
+@IF EXIST "%DEVENV%\conda-meta\history" @GOTO :ALREADY
+@ECHO Downloading Miniconda3-latest-Windows-x86_64.exe as miniconda.exe
+@powershell.exe -NoProfile -Command (new-object System.Net.WebClient).DownloadFile('https://repo.continuum.io/miniconda/Miniconda3-latest-Windows-x86_64.exe','miniconda.exe')
+@ECHO Installing miniconda to: %DEVENV%
+@start /wait "" miniconda.exe /InstallationType=JustMe /RegisterPython=0 /AddToPath=0 /S /D=%DEVENV%
+call "%DEVENV%"\Scripts\activate.bat
 conda install -y defaults::git
 @ECHO exit at this point.
 
 :ALREADY
-call "%devenv%"\Scripts\activate.bat
+call "%DEVENV%"\Scripts\activate.bat
 
 @REM Unset some variables that get in the way
 set CONDA_BAT=
@@ -36,14 +34,14 @@ set CONDA_SHLVL=
 set PYTHONPATH=
 set PYTHONHOME=
 
-@ECHO               ^>^> conda update -p "%devenv%" -yq --all
-@CALL "%devenv%\Scripts\conda" update -p "%devenv%" -yq --all
-@ECHO               ^>^> conda install -yp "%devenv%" defaults::git
-@CALL "%devenv%\Scripts\conda" install -yp "%devenv%" defaults::git
-@ECHO               ^>^> conda install -yq -p "%devenv%" --file dev/test-requirements.txt -c defaults
-@CALL "%devenv%\Scripts\conda" install -yq -p "%devenv%" --file dev/test-requirements.txt -c defaults
+@ECHO               ^>^> conda update -p "%DEVENV%" -yq --all
+@CALL "%DEVENV%\Scripts\conda" update -p "%DEVENV%" -yq --all
+@ECHO               ^>^> conda install -yp "%DEVENV%" defaults::git
+@CALL "%DEVENV%\Scripts\conda" install -yp "%DEVENV%" defaults::git
+@ECHO               ^>^> conda install -yq -p "%DEVENV%" python="%PYTHON%" pywin32 --file tests\requirements.txt -c defaults
+@CALL "%DEVENV%\Scripts\conda" install -yq -p "%DEVENV%" python="%PYTHON%" pywin32 --file tests\requirements.txt -c defaults
 
-@CALL "%devenv%\python" -m conda init --dev cmd.exe > NUL
+@CALL "%DEVENV%\python" -m conda init --dev cmd.exe > NUL
 
 :INIT_BUILD
 @CALL dev-init.bat
