--- dev/test-requirements.txt
@@ -1,23 +1,0 @@
-# run as 'conda install --file dev/test-requirements.txt -c defaults'
-# conda
-conda-package-handling
-flake8
-mock
-pexpect
-pytest!=3.7.3
-pytest-cov
-pytest-timeout
-radon
-responses
-git
-
-conda-build
-anaconda-client
-beautifulsoup4
-chardet
-conda-verify
-contextlib2
-filelock
-pycrypto
-pyflakes
-pytest-mock
