+++ setup.cfg
@@ -2,7 +2,6 @@
 minversion = 3.0
 testpaths = conda tests
 norecursedirs = .* *.egg* build dist env ve conda/_vendor/* conda_env/* test_data/*
-usefixtures = suppress_resource_warning
 addopts =
     --ignore setup.py
     --ignore conda/__main__.py
@@ -12,9 +11,13 @@ addopts =
     --ignore tests/conda_env/utils/test_uploader.py
     --ignore test_data
     --tb native
-    --strict
+    --strict-markers
     --durations 16
     --doctest-modules
+    --cov-report xml
+    --cov-report term-missing
+    --cov conda
+    --cov-append
 doctest_optionflags =
     NORMALIZE_WHITESPACE
     IGNORE_EXCEPTION_DETAIL
@@ -22,7 +25,6 @@ doctest_optionflags =
 markers =
     integration: integration tests that usually require an internet connect
     slow: slow running tests
-    installed: tests that assume conda is pre-installed on PATH
 
 
 [pycodestyle]
